Revised the plugin code to allow for easier registrations with the rtmath classes. Only a single function call and function definition need to be made. I did this all to allow silo write of shapestats as a proxy to the shapefile writer. It improves the code design.
TODOs:
- When writing a single file, reenable custom functions.
- Filename write in mesh names for silo doesn't currently work.
- Need to provide a voronoi write class.
- Shapefile extras writing seems currently broken. Need to fix to reenable visualization.


git-svn-id: file:///mnt/svn/rtmath@1273 1d52797c-35fc-48e8-8e80-2db72aa187e2

+++ rtmath/rtmath/src/shapefile.cpp
@@ -158,7 +158,7 @@ namespace rtmath {
 				latticePtsStd.conservativeResize(numPoints, 3);
 				latticePtsNorm.conservativeResize(numPoints, 3);
 				latticeIndex.conservativeResize(numPoints);
-				latticeIndex.setLinSpaced(numPoints,1,numPoints);
+				latticeIndex.setLinSpaced(numPoints,1,(int) numPoints);
 				//for (auto &extra : latticeExtras)
 				//	extra.second->conservativeResize(numPoints, 3);
 			}
